requirements.txt
openai>=1.0.0
python-dotenv>=0.21.0
scikit-learn>=1.0.0
numpy>=1.21.0
requests>=2.28.0
fastapi>=0.95.0
uvicorn>=0.22.0

fastapi
uvicorn
requests
python-dotenv